Change external mismatch check & fix to most common
fixes #117

When a developer adds a new dependency, npm/yarn will add the
newest version of the dependency. We want manypkg to fix that newest
version to the most commonly used version in the repo, since it is
unlikely that a developer adding a new dependency to a package also
intends to update the whole repo.

If the dependency ranges for a dependency all occur once, it will
fix to the highest version.